+++ src/operations/merge.js
@@ -14,6 +14,11 @@ export function merge(counters, path, position) {
     let removed = null
 
     counters = update(counters, path.slice(0, -1), (walker, state, counters) => {
+        if (index >= state.nodes.size) {
+            // Path doesn't exist.
+            return state
+        }
+
         // First remove this node from its parent, ...
         removed = state.nodes.get(index)
         state = state.deleteIn(['nodes', index])
@@ -25,6 +30,12 @@ export function merge(counters, path, position) {
 
             // performing recursive update of that node
             const prev = position === 0 ? state : state.nodes.get(position - 1)
+
+            if (!prev) {
+                // Path doesn't exist.
+                return state
+            }
+
             walker.reset(counters.values.get(prev.key))
 
             return updateTree(walker, state, position)
@@ -33,6 +44,11 @@ export function merge(counters, path, position) {
         // and finally update parent
         return updateTree(walker, state, index + 1)
     })
+
+    if (removed == null) {
+        // Path didn't exist.
+        return counters
+    }
 
     return counters.deleteIn(['values', removed.key])
 }
